refactor: Split UmeroKit into separate extension files
- Split UmeroKit.swift into separate extension files for better organization
- Created UmeroKit+Scrobbling.swift for scrobbling and authentication methods
- Created UmeroKit+Album.swift for album-related methods
- Created UmeroKit+Artist.swift for artist-related methods
- Created UmeroKit+Tag.swift for tag-related methods
- Created UmeroKit+Chart.swift for chart-related methods
- Created UmeroKit+Geo.swift for geo-related methods
- Created UmeroKit+Track.swift for track-related methods
- Created UmeroKit+User.swift for user-related methods
- Kept core struct definition and Bool extension in main UmeroKit.swift file
- Changed property access from private to internal to allow extension access
- All tests passing, SwiftLint clean
